[refactor] Styling refactor - addressing comments
