+++ modules/data-entry/src/main/frontend/src/dataHomepage/FilterComponents/BooleanFilter.jsx
@@ -26,7 +26,7 @@ import { withStyles } from 'tss-react/mui';
 import { DEFAULT_COMPARATORS, UNARY_COMPARATORS } from "./FilterComparators.jsx";
 import FilterComponentManager from "./FilterComponentManager.jsx";
 import { checkPropTypes } from "../../propTypes";
-import filterStyles from "../../questionnaire/filterStyles.jsx";
+import inputStyles from "../../questionnaire/inputStyles.jsx";
 
 const COMPARATORS = DEFAULT_COMPARATORS.slice().concat(UNARY_COMPARATORS);
 
@@ -82,7 +82,7 @@ BooleanFilter.propTypes = {
   onChangeInput: PropTypes.func
 }
 
-const StyledBooleanFilter = withStyles(BooleanFilter, filterStyles);
+const StyledBooleanFilter = withStyles(BooleanFilter, inputStyles);
 
 export default StyledBooleanFilter;
