+++ modules/data-entry/src/main/frontend/src/dataHomepage/FilterComponents/QuestionnaireFilter.jsx
@@ -27,7 +27,7 @@ import { withStyles } from 'tss-react/mui';
 import { DEFAULT_COMPARATORS } from "./FilterComparators.jsx";
 import FilterComponentManager from "./FilterComponentManager.jsx";
 import { checkPropTypes } from "../../propTypes";
-import filterStyles from "../../questionnaire/filterStyles.jsx";
+import inputStyles from "../../questionnaire/inputStyles.jsx";
 
 const COMPARATORS = DEFAULT_COMPARATORS.slice();
 
@@ -125,7 +125,7 @@ QuestionnaireFilter.propTypes = {
   onChangeInput: PropTypes.func
 }
 
-const StyledQuestionnaireFilter = withStyles(QuestionnaireFilter, filterStyles);
+const StyledQuestionnaireFilter = withStyles(QuestionnaireFilter, inputStyles);
 
 export default StyledQuestionnaireFilter;
