+++ modules/data-entry/src/main/frontend/src/dataHomepage/NewFormDialog.jsx
@@ -31,13 +31,13 @@ import { useNavigate } from "react-router";
 import { withStyles } from 'tss-react/mui';
 import { v4 as uuidv4 } from 'uuid';
 
+import tableDialogStyles from "./tableDialogStyles.jsx";
 import FormattedText from "../components/FormattedText.jsx";
 import NewItemButton from "../components/NewItemButton.jsx";
 import ResponsiveDialog from "../components/ResponsiveDialog"; // commons
 import { escapeJQL } from "../escape.jsx";
 import { fetchWithReLogin, GlobalLoginContext } from "../login/ReLoginDialog.js";
 import SubjectSelectorList, { NewSubjectDialog, parseToArray } from "../questionnaire/SubjectSelector.jsx";
-import { subjectSelectorDialogStyles } from "../questionnaire/subjectSelectorDialogStyles.jsx";
 
 const PROGRESS_SELECT_QUESTIONNAIRE = 0;
 const PROGRESS_SELECT_SUBJECT = 1;
@@ -531,4 +531,4 @@ function NewFormDialog(props) {
   )
 }
 
-export default withStyles(NewFormDialog, subjectSelectorDialogStyles);
+export default withStyles(NewFormDialog, tableDialogStyles);
