+++ modules/data-entry/src/main/frontend/src/dataHomepage/UserDashboard.jsx
@@ -28,9 +28,9 @@ import {
 import { MaterialReactTable } from "material-react-table";
 import { withStyles } from 'tss-react/mui';
 
+import dashboardStyles from "./dashboardStyles.jsx";
 import NewItemButton from "../components/NewItemButton.jsx";
 import ResponsiveDialog from "../components/ResponsiveDialog"; // commons
-import dashboardStyles from "../questionnaire/dashboardStyles.jsx";
 import { loadExtensions } from "../uiextension/extensionManager";
 
 async function getDashboardExtensions() {
