+++ modules/data-entry/src/main/frontend/src/questionnaire/ComputedQuestion.jsx
@@ -27,8 +27,8 @@ import { checkPropTypes } from "../propTypes";
 import Answer, { VALUE_POS } from "./Answer";
 import AnswerComponentManager from "./AnswerComponentManager";
 import { useFormReaderContext } from "./FormContext";
+import inputStyles from './inputStyles';
 import Question from "./Question";
-import questionStyles from './questionStyles.jsx';
 import DateTimeUtilities from "../components/DateTimeUtilities";
 import FormattedText from "../components/FormattedText";
 import { MakeRequest } from "../vocabQuery/util.jsx";
@@ -372,7 +372,7 @@ ComputedQuestion.propTypes = {
   }).isRequired
 };
 
-const StyledComputedQuestion = withStyles(ComputedQuestion, questionStyles);
+const StyledComputedQuestion = withStyles(ComputedQuestion, inputStyles);
 export default StyledComputedQuestion;
 
 AnswerComponentManager.registerAnswerComponent((definition) => {
