+++ modules/data-entry/src/main/frontend/src/questionnaire/FormEntry.jsx
@@ -55,6 +55,8 @@ export const SECTION_TYPES = ["cards:Section"];
 export const INFO_TYPES = ["cards:Information"];
 export const ENTRY_TYPES = QUESTION_TYPES.concat(SECTION_TYPES).concat(INFO_TYPES);
 
+let hiddenStyle = { display: "none" };
+
 /**
  * Method responsible for displaying a question from the questionnaire, along with its answer(s).
  *
@@ -70,7 +72,6 @@ let displayQuestion = (
   path,
   existingAnswer,
   key,
-  classes,
   onAddedAnswerPath,
   sectionAnswersState,
   onChange,
@@ -92,15 +93,15 @@ let displayQuestion = (
   // This variable must start with an upper case letter so that React treats it as a component
   const QuestionDisplay = AnswerComponentManager.getAnswerComponent(questionDefinition);
 
-  let gridClasses = [];
+  let gridStyle = {};
   let displayMode = questionDefinition.displayMode;
   if (pageActive === false || displayMode == 'hidden' || (isSummary && displayMode !== "summary") || (!isSummary && displayMode === "summary")) {
-    gridClasses.push(classes.hiddenQuestion);
+    gridStyle = hiddenStyle;
   }
 
   // component will either render the default question display, or a list of questions/answers from the form (used for subjects)
   return (
-    <Grid key={key} className={gridClasses.join(" ")} {...gridProps}>
+    <Grid key={key} sx={gridStyle} {...gridProps}>
       <QuestionDisplay
         questionDefinition={questionDefinition}
         existingAnswer={existingQuestionAnswer}
@@ -176,7 +177,7 @@ let displaySection = (
   );
 }
 
-let displayInformation = (infoDefinition, key, classes, pageActive, isEdit, gridProps) => {
+let displayInformation = (infoDefinition, key, pageActive, isEdit, gridProps) => {
   let isVisible = (
     (!infoDefinition.formMode || infoDefinition.formMode == "edit") && isEdit
     || infoDefinition.formMode == "view" && !isEdit
@@ -198,7 +199,6 @@ let displayInformation = (infoDefinition, key, classes, pageActive, isEdit, grid
  * @param {string} path the path to the parent of the question
  * @param {Object} existingAnswer form data that may include answers already submitted for this component
  * @param {string} key the node name of the question definition JCR node
- * @param {Object} classes style classes
  * @returns a React component that renders the matrix section
  */
 let displayMatrix = (
@@ -206,7 +206,6 @@ let displayMatrix = (
   path,
   existingAnswer,
   key,
-  classes,
   pageActive,
   isEdit,
   contentOffset,
@@ -230,13 +229,13 @@ let displayMatrix = (
     return null;
   }
 
-  let gridClasses = [];
+  let gridStyle = {};
   if (pageActive === false || sectionDefinition.displayMode == 'hidden') {
-    gridClasses.push(classes.hiddenQuestion);
+    gridStyle = hiddenStyle;
   }
 
   return (
-    <Grid key={key} className={gridClasses.join(" ")} {...gridProps}>
+    <Grid key={key} sx={gridStyle} {...gridProps}>
       <QuestionMatrix
         sectionDefinition={sectionDefinition}
         existingSectionAnswer={existingSectionAnswer}
@@ -258,12 +257,10 @@ let displayMatrix = (
  * @param {int} depth the section nesting depth
  * @param {Object} existingAnswers form data that may include answers already submitted for this component
  * @param {string} key the node name of the section definition JCR node
- * @param {Object} classes style classes
  * @returns a React component that renders the section
  */
 export default function FormEntry(props) {
   let {
-    classes,
     entryDefinition,
     path,
     depth,
@@ -290,7 +287,6 @@ export default function FormEntry(props) {
       path,
       existingAnswers,
       keyProp,
-      classes,
       onAddedAnswerPath,
       sectionAnswersState,
       onChange,
@@ -308,7 +304,6 @@ export default function FormEntry(props) {
         path,
         existingAnswers,
         keyProp,
-        classes,
         pageActive,
         isEdit,
         contentOffset,
@@ -332,6 +327,6 @@ export default function FormEntry(props) {
       );
     }
   } else if (INFO_TYPES.includes(entryDefinition["jcr:primaryType"])) {
-    return displayInformation(entryDefinition, keyProp, classes, pageActive, isEdit, gridProps);
+    return displayInformation(entryDefinition, keyProp, pageActive, isEdit, gridProps);
   }
 }
