+++ modules/data-entry/src/main/frontend/src/questionnaire/NumberQuestion.jsx
@@ -34,9 +34,9 @@ import Answer from "./Answer";
 import AnswerComponentManager from "./AnswerComponentManager";
 import AnswerInstructions from "./AnswerInstructions";
 import { useFormReaderContext } from "./FormContext";
+import inputStyles from "./inputStyles";
 import MultipleChoice from "./MultipleChoice";
 import Question from "./Question";
-import questionStyles from "./questionStyles.jsx";
 import FormattedText from "../components/FormattedText";
 
 /** Conversion between the `dataType` setting in the question definition and the corresponding primary node type of the `Answer` node for that question. */
@@ -578,7 +578,7 @@ NumberQuestion.propTypes = {
   isRange: PropTypes.bool,
 };
 
-const StyledNumberQuestion = withStyles(NumberQuestion, questionStyles);
+const StyledNumberQuestion = withStyles(NumberQuestion, inputStyles);
 export default StyledNumberQuestion;
 
 AnswerComponentManager.registerAnswerComponent((questionDefinition) => {
