+++ modules/data-entry/src/main/frontend/src/questionnaire/Section.jsx
@@ -309,7 +309,6 @@ function Section(props) {
                           depth={depth+1}
                           existingAnswers={existingSectionAnswer}
                           keyProp={key}
-                          classes={classes}
                           onChange={onChange}
                           isEdit={isEdit}
                           isSummary={isSummary}
