+++ modules/data-entry/src/main/frontend/src/questionnaire/SelectableAreaQuestion.jsx
@@ -23,14 +23,12 @@ import { Alert, Checkbox, FormControlLabel, Typography } from "@mui/material";
 import { useTheme, alpha } from '@mui/material/styles';
 import Tooltip from "@mui/material/Tooltip";
 import PropTypes from "prop-types";
-import { withStyles } from 'tss-react/mui';
 
 import { checkPropTypes } from "../propTypes";
 import Answer, { LABEL_POS, VALUE_POS } from "./Answer";
 import AnswerComponentManager from "./AnswerComponentManager";
 import { useFormReaderContext } from "./FormContext";
 import Question from "./Question";
-import questionMatrixStyles from "./questionMatrixStyles.jsx";
 import FormattedText from "../components/FormattedText.jsx";
 
 // Component that renders an image with clickable areas based on the available
@@ -53,7 +51,7 @@ import FormattedText from "../components/FormattedText.jsx";
 //    />
 function SelectableAreaQuestion(props) {
   checkPropTypes(SelectableAreaQuestion, props);
-  let { classes, existingAnswer, questionName, questionDefinition, pageActive, isEdit, ...rest } = props;
+  let { existingAnswer, questionName, questionDefinition, pageActive, isEdit, ...rest } = props;
   let { variant, maxAnswers } = { ...props.questionDefinition, ...props };
 
   const [ map, setMap ] = useState(null);
@@ -358,7 +356,7 @@ function SelectableAreaQuestion(props) {
               label={notApplicableOption.label || notApplicableOption.value}
               value={notApplicableOption.value}
             />
-            <FormattedText className={classes.selectionDescription} variant="caption" color="textSecondary">
+            <FormattedText sx={{ pl: 4 }} variant="caption" color="textSecondary">
               {notApplicableOption.help}
             </FormattedText>
           </>
@@ -414,11 +412,10 @@ SelectableAreaQuestion.propTypes = {
   maxAnswers: PropTypes.number
 };
 
-const StyledSelectableAreaQuestion = withStyles(SelectableAreaQuestion, questionMatrixStyles);
-export default StyledSelectableAreaQuestion;
+export default SelectableAreaQuestion;
 
 AnswerComponentManager.registerAnswerComponent((questionDefinition) => {
   if (questionDefinition.dataType === "selectableArea") {
-    return [StyledSelectableAreaQuestion, 50];
+    return [SelectableAreaQuestion, 50];
   }
 });
