+++ modules/data-entry/src/main/frontend/src/questionnaire/multipleChoiceStyles.jsx
@@ -1,0 +1,85 @@
+//
+//  Licensed to the Apache Software Foundation (ASF) under one
+//  or more contributor license agreements.  See the NOTICE file
+//  distributed with this work for additional information
+//  regarding copyright ownership.  The ASF licenses this file
+//  to you under the Apache License, Version 2.0 (the
+//  "License"); you may not use this file except in compliance
+//  with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing,
+//  software distributed under the License is distributed on an
+//  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+//  KIND, either express or implied.  See the License for the
+//  specific language governing permissions and limitations
+//  under the License.
+//
+
+import inputStyles from "./inputStyles.jsx";
+
+const multipleChoiceStyles = theme => ({
+  ...inputStyles(theme),
+  ghostListItem: {
+    paddingTop: 0,
+    alignItems: "flex-start",
+  },
+  searchWrapper: {
+    margin: 0,
+    position: 'relative',
+    display: 'inline-block',
+    padding: theme.spacing(.5, 0, 0, 0),
+  },
+  nestedInput: {
+    minWidth: "218px !important",
+    marginLeft: theme.spacing(-2.5),
+  },
+  selectMultiValues: {
+    whiteSpace: "normal",
+    "& .MuiChip-root" : {
+      margin: "1px",
+    },
+  },
+  optionsList: {
+    padding: theme.spacing(0),
+  },
+  deleteButton: {
+    padding: theme.spacing(1,0),
+    margin: theme.spacing(-1,0,-1,-1.5),
+    fontSize: "10px",
+    minWidth: "42px",
+    "& + div" : {
+      marginRight: theme.spacing(2),
+    },
+  },
+  compactLayout : {
+    "& .MuiList-root" : {
+      [theme.breakpoints.up('sm')]: {
+        display: "inline-block",
+      },
+    },
+    "& .MuiListItem-root" : {
+      [theme.breakpoints.up('sm')]: {
+        display: "inline-flex",
+        width: "auto",
+      },
+    },
+    "& .MuiListItem-root > div:first-of-type > .MuiTextField-root" : {
+      [theme.breakpoints.up('sm')]: {
+        minWidth: "100px",
+        marginTop: theme.spacing(-1.5),
+      },
+    },
+  },
+  selectionChild: {
+    cursor: "pointer",
+    flexWrap: "wrap",
+  },
+  selectionDescription: {
+    flexBasis: "100%",
+    paddingLeft: theme.spacing(4)
+  },
+});
+
+export default multipleChoiceStyles;
