--- modules/data-entry/src/main/frontend/src/questionnaire/stickySectionStyles.jsx
@@ -1,45 +1,0 @@
-//
-//  Licensed to the Apache Software Foundation (ASF) under one
-//  or more contributor license agreements.  See the NOTICE file
-//  distributed with this work for additional information
-//  regarding copyright ownership.  The ASF licenses this file
-//  to you under the Apache License, Version 2.0 (the
-//  "License"); you may not use this file except in compliance
-//  with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing,
-//  software distributed under the License is distributed on an
-//  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-//  KIND, either express or implied.  See the License for the
-//  specific language governing permissions and limitations
-//  under the License.
-//
-
-import { grey } from '@mui/material/colors';
-
-// Shared factory for sticky header/footer questionnaire sections
-export const stickySectionStyles = (theme, position = "top") => ({
-  "&.cards-edit-section" : {
-    position: "sticky",
-    ...(position === "top"
-      ? { top: 0, zIndex: 2, paddingTop: 0 }
-      : { bottom: 0, paddingBottom: "0 !important" }),
-  },
-  "& > .MuiCollapse-wrapper" : {
-    border: "1px solid " + theme.palette.primary.light,
-  },
-  "& .MuiGrid-root:not(:first-of-type)": {
-    paddingTop: 0,
-  },
-  "& .MuiGrid-root:not(:last-child)": {
-    paddingBottom: 0,
-  },
-  "& .MuiGrid-root:not(.MuiCollapse-container) > *": {
-    background: grey[100],
-  },
-  "& .MuiCard-root" : {
-    borderColor: "transparent",
-  },
-});
